Fix Next.js client/server component conflict for Vercel deployment
Split BotDetailPage into server component (params handling) and
BotDashboardClient (Zustand hooks). Fixes 'use client' + async function error.
